add smtp domain to get rid of open relay, change default port of dns
